ticket_generate/requirements.txt
requests==2.26.0
python-dotenv==0.17.1
pypng==0.0.21
PyQRCode==1.2.1
pymongo==3.10.1
Jinja2==2.11.3